pyproject.toml
[project]
name = "mlx-serve"
version = "0.1.0"
description = "Local inference server for Apple Silicon that hot-swaps MLX models (LLM, vision, embeddings, TTS, STT) via OpenAI-compatible API"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.11"
authors = [
    { name = "mlx-serve contributors" },
]
keywords = [
    "mlx", "apple-silicon", "llm", "openai", "openai-api",
    "inference-server", "local-llm", "local-inference",
    "model-manager", "model-serving", "fastapi", "macos", "metal",
    "text-to-speech", "tts", "speech-to-text", "stt",
    "embeddings", "vision-language-model",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: No Input/Output (Daemon)",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Typing :: Typed",
]
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "psutil>=5.9",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
text = ["mlx-lm>=0.14"]
vision = ["mlx-vlm>=0.0.9"]
tts = ["mlx-audio>=0.4.0"]
stt = ["mlx-whisper>=0.4.0"]
embeddings = ["mlx-embeddings>=0.1.0"]
all = [
    "mlx-lm>=0.14",
    "mlx-vlm>=0.0.9",
    "mlx-audio>=0.4.0",
    "mlx-embeddings>=0.1.0",
    "mlx-whisper>=0.4.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.4",
]

[project.scripts]
mlx-serve = "mlx_serve.cli:main"

[project.urls]
Homepage = "https://github.com/raspoli/mlx-serve"
Repository = "https://github.com/raspoli/mlx-serve"
Issues = "https://github.com/raspoli/mlx-serve/issues"
Changelog = "https://github.com/raspoli/mlx-serve/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mlx_serve"]

[tool.hatch.version]
path = "src/mlx_serve/__init__.py"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
extend-select = [
    "I",    # isort — import sorting
    "UP",   # pyupgrade — modernize syntax
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "PTH",  # flake8-use-pathlib
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]
"__init__.py" = ["F401"]
